+++ app/assets/javascripts/backend/application.js
@@ -14,6 +14,9 @@
 //= require jquery_ujs
 //= require jquery-ui/autocomplete
 //= require jquery-ui/sortable
+//= require backend/tether.min
 //= require backend/bootstrap
 //= require ckeditor/init
 //= require backend/general
+//= require backend/bootstrap-datepicker.min
+//= require backend/datepickers