+++ app/assets/javascripts/backend/general.js
@@ -1,4 +1,4 @@
-var general = {
+var general = general || {
   init: function() {
     $('form:not(.no-focus) input:visible:not(.no-focus):first').focus();
   }